factored out data binning functionality into multiple classes
